[Fix] #7 - 폴더링 변경
