--- AraBook/AraBook/Presentation/RecordsList/Model/RecordListModel.swift
@@ -1,42 +1,0 @@
-//
-//  RecordListModel.swift
-//  AraBook
-//
-//  Created by KJ on 9/27/24.
-//
-
-import Foundation
-
-struct RecordListModel {
-    let id: Int
-    let title: String
-    let totalReadDate: String
-    let readDate: String
-    let backgroundColor: String
-}
-
-extension RecordListModel {
-    
-    static func dummy() -> [RecordListModel] {
-        return [
-            RecordListModel(id: 0,
-                                   title: "퀸의 대각선",
-                                   totalReadDate : "총 7일 동안 읽었어요!",
-                                   readDate: "23. 11. 25 ~ 24. 09. 20",
-                            backgroundColor: "yellow"
-            ),
-            RecordListModel(id: 2,
-                                   title: "퀸의 대각선",
-                                   totalReadDate : "총 7일 동안 읽었어요!",
-                                   readDate: "23. 11. 25 ~ 24. 09. 20",
-                            backgroundColor: "green"
-            ),
-            RecordListModel(id: 3,
-                                   title: "퀸의 대각선",
-                                   totalReadDate : "총 7일 동안 읽었어요!",
-                                   readDate: "23. 11. 25 ~ 24. 09. 20",
-                            backgroundColor: "orange"
-            )
-        ]
-    }
-}
